+++ app/src/main/java/jp/matsuura/pokemon/androidapp/ui/Destination.kt
@@ -2,7 +2,5 @@ package jp.matsuura.pokemon.androidapp.ui
 
 sealed class Destination (val route: String) {
     object Home : Destination(route = "home")
-    object Detail : Destination(route = "detail") {
-
-    }
+    object Detail : Destination(route = "detail")
 }
