--- app/src/main/java/jp/matsuura/pokemon/androidapp/ui/detail/DetailScreenEvent.kt
@@ -1,6 +1,0 @@
-package jp.matsuura.pokemon.androidapp.ui.detail
-
-sealed interface DetailScreenEvent {
-    object NetworkError: DetailScreenEvent
-    data class UnknownError(val error: Throwable): DetailScreenEvent
-}
