--- app/src/main/java/jp/matsuura/pokemon/androidapp/ui/home/HomeScreenState.kt
@@ -1,15 +1,0 @@
-package jp.matsuura.pokemon.androidapp.ui.home
-
-import jp.matsuura.pokemon.model.PokemonModel
-
-data class HomeScreenState(
-    val isLoading: Boolean,
-    val pokemonList: List<PokemonModel>,
-) {
-    companion object {
-        fun initValue() = HomeScreenState(
-            isLoading = false,
-            pokemonList = emptyList(),
-        )
-    }
-}
