+++ settings.gradle.kts
@@ -19,3 +19,6 @@ include(":data:api")
 include(":data:repository")
 include(":domain")
 include(":model")
+include(":ui:home")
+include(":ui:common")
+include(":ui:detail")
