+++ ui/common/src/main/java/jp/matsuura/pokemon/common/ProgressIndicator.kt
@@ -1,0 +1,26 @@
+package jp.matsuura.pokemon.common
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+
+@Composable
+fun ProgressIndicator() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Color.DarkGray),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        CircularProgressIndicator(
+            modifier = Modifier
+                .wrapContentHeight()
+                .wrapContentHeight()
+        )
+    }
+}
