+++ ui/detail/.gitignore
@@ -1,0 +1,1 @@
+/build
