+++ ui/detail/src/main/java/jp/matsuura/pokemon/detail/DetailScreenEvent.kt
@@ -1,0 +1,6 @@
+package jp.matsuura.pokemon.detail
+
+sealed interface DetailScreenEvent {
+    object NetworkError: DetailScreenEvent
+    data class UnknownError(val error: Throwable): DetailScreenEvent
+}
