+++ ui/detail/src/main/java/jp/matsuura/pokemon/detail/DetailScreenState.kt
@@ -1,0 +1,15 @@
+package jp.matsuura.pokemon.detail
+
+import jp.matsuura.pokemon.model.PokemonDetailModel
+
+data class DetailScreenState(
+    val isProgressVisible: Boolean,
+    val pokemonInfo: PokemonDetailModel?,
+) {
+    companion object {
+        fun initValue() = DetailScreenState(
+            isProgressVisible = false,
+            pokemonInfo = null,
+        )
+    }
+}
