+++ ui/home/src/main/java/jp/matsuura/pokemon/home/HomeScreen.kt
@@ -1,0 +1,128 @@
+package jp.matsuura.pokemon.home
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import coil.compose.AsyncImage
+import jp.matsuura.pokemon.common.ProgressIndicator
+import jp.matsuura.pokemon.model.PokemonModel
+
+@Composable
+fun HomeScreen(
+    viewModel: HomeViewModel = hiltViewModel(),
+    onCardItemClicked: (Int) -> Unit,
+) {
+    val state by viewModel.uiState.collectAsStateWithLifecycle()
+    HomeScreen(
+        state = state,
+        onCardItemClicked = { pokemonId ->
+            // FIXME: notify the event to ViewModel.
+            onCardItemClicked(pokemonId.toInt())
+        }
+    )
+    // TODO: handle the one-shot event.
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun HomeScreen(
+    state: HomeScreenState,
+    onCardItemClicked: (String) -> Unit,
+) {
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = {
+                    Text(text = "Pokemon Library")
+                },
+            )
+        },
+    ) {
+        Box(
+            modifier = Modifier.padding(it),
+        ) {
+            if (state.isLoading) ProgressIndicator()
+            PokemonListItem(
+                pokemonList = state.pokemonList,
+                onCardItemClicked = onCardItemClicked,
+            )
+        }
+    }
+}
+
+@Composable
+fun PokemonListItem(
+    pokemonList: List<PokemonModel>,
+    onCardItemClicked: (String) -> Unit,
+) {
+    LazyVerticalGrid(
+        columns = GridCells.Fixed(2),
+    ) {
+        items(pokemonList) { pokemon ->
+            PokemonItem(
+                pokemon = pokemon,
+                onCardItemClicked = onCardItemClicked,
+            )
+        }
+    }
+}
+
+@Composable
+fun PokemonItem(
+    pokemon: PokemonModel,
+    onCardItemClicked: (String) -> Unit,
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(top = 8.dp, start = 16.dp, end = 16.dp, bottom = 16.dp)
+    ) {
+        // FIXME: fix the ripple effect.
+        Card(
+            shape = RoundedCornerShape(20.dp),
+            colors = CardDefaults.cardColors(
+                containerColor = Color.White,
+            ),
+            elevation = CardDefaults.cardElevation(
+                defaultElevation = 8.dp
+            ),
+            modifier = Modifier.clickable {
+                onCardItemClicked(pokemon.id)
+            }
+        ) {
+            AsyncImage(
+                model = pokemon.imageUrl,
+                contentDescription = null,
+                modifier = Modifier.padding(top = 16.dp, start = 24.dp, end = 24.dp)
+            )
+            Text(
+                text = "No.${pokemon.id}",
+                fontSize = 14.sp,
+                textAlign = TextAlign.Center,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 8.dp),
+            )
+            Text(
+                text = pokemon.jaName,
+                fontSize = 14.sp,
+                textAlign = TextAlign.Center,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 4.dp, bottom = 16.dp),
+            )
+        }
+    }
+}
