+++ ui/home/src/main/java/jp/matsuura/pokemon/home/HomeScreenEvent.kt
@@ -1,0 +1,7 @@
+package jp.matsuura.pokemon.home
+
+sealed interface HomeScreenEvent {
+    object NetworkError: HomeScreenEvent
+    data class UnknownError(val error: Throwable): HomeScreenEvent
+    data class NavigateToDetail(val pokemonId: Int): HomeScreenEvent
+}
